# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Receipt Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CurrentUser {
  id: string;
  /** @nullable */
  email?: string | null;
  isAdmin: boolean;
}

export interface AdminUser {
  id: string;
  /** @nullable */
  email?: string | null;
  isAdmin: boolean;
  createdAt: string;
  storeCount: number;
  itemCount: number;
  receiptCount: number;
  totalSpend: number;
}

export interface AdminReceiptSummary {
  id: number;
  storeName: string;
  total: number;
  purchasedAt: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface AdminUserReceipts {
  userId: string;
  /** @nullable */
  email?: string | null;
  receipts: AdminReceiptSummary[];
}

export interface Store {
  id: number;
  name: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  openTimes?: string | null;
  deliveryAvailable: boolean;
  /** @nullable */
  deliveryFee?: number | null;
  /** @nullable */
  minimumOrderAmount?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface StoreInput {
  name: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  openTimes?: string | null;
  deliveryAvailable?: boolean;
  /** @nullable */
  deliveryFee?: number | null;
  /** @nullable */
  minimumOrderAmount?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface StoreUpdate {
  name?: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  openTimes?: string | null;
  deliveryAvailable?: boolean;
  /** @nullable */
  deliveryFee?: number | null;
  /** @nullable */
  minimumOrderAmount?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface Item {
  id: number;
  name: string;
  /**
     * Emoji icon representing the item
     * @nullable
     */
  icon?: string | null;
  /** @nullable */
  notes?: string | null;
  purchaseCount: number;
  createdAt: string;
}

export interface ItemInput {
  name: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface ItemUpdate {
  name?: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Receipt {
  id: number;
  storeId: number;
  storeName: string;
  purchasedAt: string;
  total: number;
  /** @nullable */
  totalBeforeTax?: number | null;
  /** @nullable */
  imageUri?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ReceiptInput {
  storeId: number;
  purchasedAt: string;
  total: number;
  /** @nullable */
  imageUri?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface LineItem {
  id: number;
  receiptId: number;
  itemId: number;
  itemName: string;
  /** @nullable */
  icon?: string | null;
  price: number;
  quantity: number;
  createdAt: string;
}

export interface LineItemInput {
  itemId: number;
  price: number;
  quantity: number;
}

export interface LineItemUpdate {
  price?: number;
  quantity?: number;
}

export interface ReceiptDetail {
  id: number;
  storeId: number;
  storeName: string;
  purchasedAt: string;
  total: number;
  /** @nullable */
  totalBeforeTax?: number | null;
  /** @nullable */
  imageUri?: string | null;
  /** @nullable */
  notes?: string | null;
  lineItems: LineItem[];
  createdAt: string;
}

export type ManualEntryInputLineItemsItem = {
  name: string;
  price: number;
  quantity: number;
};

export interface ManualEntryInput {
  storeName: string;
  /** @nullable */
  storeAddress?: string | null;
  /** @nullable */
  storePhone?: string | null;
  /** @nullable */
  storeOpenTimes?: string | null;
  /** ISO 8601 date-time string */
  purchasedAt: string;
  /** Total after tax */
  total: number;
  /** @nullable */
  totalBeforeTax?: number | null;
  /** @nullable */
  notes?: string | null;
  lineItems: ManualEntryInputLineItemsItem[];
}

export interface StoreVisitItem {
  itemName: string;
  price: number;
  quantity: number;
}

export interface StoreVisit {
  receiptId: number;
  purchasedAt: string;
  items: StoreVisitItem[];
}

export interface StoreVisitsReport {
  storeId: number;
  storeName: string;
  visits: StoreVisit[];
  uniqueItems: string[];
}

export interface ItemHistoryEntry {
  receiptId: number;
  purchasedAt: string;
  storeName: string;
  price: number;
  quantity: number;
}

export interface ItemHistoryReport {
  itemId: number;
  itemName: string;
  /** @nullable */
  icon?: string | null;
  purchaseCount: number;
  averagePrice: number;
  lowestPrice: number;
  highestPrice: number;
  /** @nullable */
  daysSinceLastPurchase?: number | null;
  /** @nullable */
  lastPurchasedAt?: string | null;
  /** @nullable */
  ranOutAt?: string | null;
  history: ItemHistoryEntry[];
}

export interface RanOutResponse {
  ranOutAt: string;
  /** @nullable */
  daysSinceLastPurchase?: number | null;
}

export interface WeeklySpend {
  weekStart: string;
  weekEnd: string;
  total: number;
  isHigh: boolean;
  isLow: boolean;
  receiptCount?: number;
}

export interface SpendAnalytics {
  weeks: WeeklySpend[];
  average: number;
  totalSpend: number;
  weeklyAverage: number;
}

export interface PricePoint {
  date: string;
  price: number;
  storeId: number;
  storeName: string;
  receiptId: number;
}

export interface ItemPriceHistory {
  itemId: number;
  itemName: string;
  /** @nullable */
  icon?: string | null;
  averagePrice: number;
  lowestPrice: number;
  highestPrice: number;
  lowestPriceStoreName: string;
  pricePoints: PricePoint[];
}

export interface StoreSummary {
  storeId: number;
  storeName: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  openTimes?: string | null;
  receiptCount: number;
  totalSpend: number;
  averageReceiptTotal: number;
  deliveryAvailable: boolean;
  /** @nullable */
  deliveryFee?: number | null;
  /** @nullable */
  minimumOrderAmount?: number | null;
  /** @nullable */
  deliveryCostBenefitNote?: string | null;
}

export interface ShoppingListItem {
  itemId: number;
  itemName: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  notes?: string | null;
  purchaseCount: number;
  averagePrice: number;
  lowestPrice: number;
  lowestPriceStoreName: string;
  isRecurring: boolean;
  /** @nullable */
  daysSinceLastPurchase?: number | null;
  /** @nullable */
  lastPurchasedAt?: string | null;
  /** @nullable */
  ranOutAt?: string | null;
}

export interface ShoppingList {
  recurring: ShoppingListItem[];
  oneOff: ShoppingListItem[];
}

export interface ParsedReceiptLineItem {
  name: string;
  /** @nullable */
  icon?: string | null;
  price: number;
  quantity: number;
  nameUncertain?: boolean;
  priceUncertain?: boolean;
}

export interface ParsedReceipt {
  storeName: string;
  storeNameUncertain?: boolean;
  purchasedAt: string;
  dateUncertain?: boolean;
  total: number;
  totalUncertain?: boolean;
  lineItems: ParsedReceiptLineItem[];
}

export interface DaySpend {
  /** ISO date string YYYY-MM-DD */
  date: string;
  total: number;
  receiptCount: number;
}

export interface ParseReceiptInput {
  imageBase64: string;
}

export interface ParsePdfInput {
  pdfBase64: string;
}

export interface ReceiptBounds {
  /** Left edge as fraction of image width (0–1) */
  x: number;
  /** Top edge as fraction of image height (0–1) */
  y: number;
  /** Width as fraction of image width (0–1) */
  width: number;
  /** Height as fraction of image height (0–1) */
  height: number;
}