# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Receipt Tracker API
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List all stores
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createStore
      tags: [stores]
      summary: Create a store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreInput"
      responses:
        "201":
          description: Created store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /stores/{id}:
    get:
      operationId: getStore
      tags: [stores]
      summary: Get a store by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "404":
          description: Not found
    patch:
      operationId: updateStore
      tags: [stores]
      summary: Update a store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreUpdate"
      responses:
        "200":
          description: Updated store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
    delete:
      operationId: deleteStore
      tags: [stores]
      summary: Delete a store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Store:
      type: object
      required: [id, name, deliveryAvailable, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        openTimes:
          type: ["string", "null"]
        deliveryAvailable:
          type: boolean
        deliveryFee:
          type: ["number", "null"]
        minimumOrderAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    StoreInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        openTimes:
          type: ["string", "null"]
        deliveryAvailable:
          type: boolean
        deliveryFee:
          type: ["number", "null"]
        minimumOrderAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    StoreUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        openTimes:
          type: ["string", "null"]
        deliveryAvailable:
          type: boolean
        deliveryFee:
          type: ["number", "null"]
        minimumOrderAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
...